Improve copy and attempt to clarify mechanism
@ko1 I tried to make this a bit easier to understand and I'm still not certain 
that I've actually improved the paragraph about what I called "premature promotion".

The way I understood it from your explanation, if there are many young objects it's 
more likely for them to be promoted early than when there are fewer young objects.

I didn't understand your explanation for the cause of this phenomenon: the part where 
you say "because object headers are written and marked as dirty page". Could you 
try to explain this more clearly maybe? I'm not sure what the "dirty page" part means 
exactly. It seems like this is an important part of your explanation.

By reading https://bugs.ruby-lang.org/issues/10137 I can tell that the object age can be 
between 0 and 3, so 0, 1, 2, or 3. I sadly don't understand how Copy on Write is involved with 
this issue by reading your explanation.

Anyway, thank you for this gem and I couldn't find a definitive answer for this on bugs.ruby-lang.org 
but is this patch included by default in Ruby 2.3? Has anyone considered backporting it to 
Ruby 2.2 or is it out of the question?